+++ FruitsGraphQL/Navigation/AppTransition.swift
@@ -1,33 +1,22 @@
 //
-//  Router.swift
+//  AppTransition.swift
 //  FruitsGraphQL
 //
-//  Created by Robert on 11/22/22.
+//  Created by Robert on 11/23/22.
 //
 
-import UIKit
-
-protocol Router: AnyObject {
-    associatedtype Route
-    var navigationController: UINavigationController { get set }
-    func exit()
-    func present(route: Route)
-}
-
-protocol AppRouter: Router where Route == AppTransition {
-    
-}
+import Foundation
 
 enum AppTransition {
     case mainList
-    
+
     func coordinaorFor<R: AppRouter>(appRouter: R) -> Coordinator {
         switch self {
         case .mainList:
             return MainListCoordinator(router: appRouter)
         }
     }
-    
+
     var transitionName: String {
         switch self {
         case .mainList:
@@ -35,5 +24,3 @@ enum AppTransition {
         }
     }
 }
-
-
